feat(dashboard/settings): show real backend URLs (closes #252)
Replaces the misleading "Backend URL: http://localhost:8080" placeholder
on /settings with three real rows: Proxy public URL, Dashboard-to-Proxy
URL, and Upstream URL.

- Convert page.tsx to a server component that reads LLMTRACE_PROXY_URL
  and forwards it as a prop to a new client component.
- Add SettingsClient with three UrlRow entries plus a copy button for
  the proxy URLs; degrades to "Not configured" (warning) when the env
  var is empty and "Unavailable" when the live config fetch fails.
- Upstream URL is sourced from upstream_url in GET /api/proxy/config/live
  (already fetched for the "Current Configuration" card).
- Update env vars table: list LLMTRACE_PROXY_URL and drop the stale
  http://localhost:8080 default that was never used in Basilica.
- Preserve health/status badge, Current Configuration card, and the
  embedded Swagger UI section.
- Playwright smoke tests covering the three-row render and the
  Unavailable degradation path.
